requirements.txt
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.7
django-rest-swagger==2.2.0
djangorestframework==3.9.2
djangorestframework-simplejwt==4.1.0
djangorestframework-stubs==0.4.2
idna==2.8
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
mysqlclient==1.4.2
openapi-codec==1.3.2
PyJWT==1.7.1
pytz==2018.9
requests==2.21.0
simplejson==3.16.0
uritemplate==3.0.0
urllib3==1.24.1
